pyproject.toml
[project]
name = "ubuntu-cycle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cloudinary>=1.44.0",
    "fastapi>=0.115.12",
    "jinja2>=3.1.4",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.12",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.34.2",
]
